tercer

+++ src/main/java/com/gestionEnvios/gestionEnvios/Repositorio/ClienteRepo.java
@@ -3,5 +3,8 @@ package com.gestionEnvios.gestionEnvios.Repositorio;
 import com.gestionEnvios.gestionEnvios.Entidades.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepo extends JpaRepository <Cliente,Long> {
+import java.util.Optional;
+
+public interface ClienteRepo extends JpaRepository <Cliente,Integer> {
+
 }
